refactor: move DeadValueCheck to own file
